+++ src/Reader/FilterableDataInterface.php
@@ -20,12 +20,19 @@ interface FilterableDataInterface
     /**
      * Returns new instance with data reading criteria set.
      *
-     * @param FilterInterface $filter Data reading criteria.
+     * @param ?FilterInterface $filter Data reading criteria.
      *
      * @return static New instance.
      * @psalm-return $this
      */
-    public function withFilter(FilterInterface $filter): static;
+    public function withFilter(?FilterInterface $filter): static;
+
+    /**
+     * Get current data reading criteria.
+     *
+     * @return FilterInterface|null Data reading criteria.
+     */
+    public function getFilter(): ?FilterInterface;
 
     /**
      * Returns new instance with additional handlers set.